[ci] inline display_ec2 script
This is purely for forward compatibility reasons. For actions and
workflows, we can pin them to a specific hash in our repo. Doing so
makes the migration smoother, because it means that to do CI we don't
need to reference the local checkout at all (and thus older PRs will not
break even if they are based on a commit that doesn't have the new
workflow files).

The only actually new file we reference is `display_ec2_info.sh`. So for
now, inline it. This incurs a little duplication but is worth it.

ghstack-source-id: 3a3456c0fa07ff1de2d0cd73cf273535cbff44bc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74355